fix issue with programNames search term, move effect and methods to reusable hook
